Add tests for Arguments
Adding #required? makes sense in this case, as it is much easier to
understand

argument.required?

than

not argument.optional?

Generally, I don’t like having mutually exclusive boolean tests like
this, but there’s a difference between optional/required and started/not
started in that optional/required are antonyms.