Allow Ame::Class.namespace to take a description
We need a way to describe an Ame::Class.  Overloading .namespace seemed
like the natural choice.  Perhaps we should have kept the old semantics
as well, checking for whether a name was passed as a String.  I think
this is a good combination, however, as you’ll seldem have to explicitly
set the namespace.