Move help generation to separate module
The question is how Help::Console will be implemented in the end.  Will
it be a class?  How do we deal with output?  Additional methods for
Method, Methods, Class, and Classes will require a name for the program,
as it is needed for the “usage” line.  Perhaps this should be a class,
after all, where the name of the program can be set in the initializer.