Methods that lack a description should cause error
The old behavior of silently ignoring methods that lacked a description,
assuming that they were public methods that simply shouldn’t have an
Ame::Method associated with them could cause a lot of user frustration,
as forgetting to describe the method would result in the method not
appearing.  We assume that any methods that are made public should have
an Ame::Method associated with them.