Don’t use Class#inherited
Class#inherited is called as soon as the class is created when using the
standard syntax, but only after the block executes when using Class#new.
For example,

class A
  def self.inherited(subclass)
    puts "#{self} was inherited by #{subclass}"
  end
end

class B < A
  puts "Defining #{self}"
end

outputs

A was inherited by B
Defining B

whereas

Class.new(A) do
  puts "Defining #{self}"
end

outputs

Defining #<Class:0x…>
A was inherited by #<Class:0x…>

Since at least #initialize has to be defined when we’re inheriting – for
its description – we can’t use Class#inherited.  Instead, move the
Dispatch creation to Module#method_added for :initialize.

Please note that the definitions of Class#process and Class#call must
come before the definition of Class.method_added, as they’re not meant
to be defined as Methods.