Parse combined short options correctly
The old code would turn “-ab” into “-a -b”, which would, if “-a” took an
argument, result in rather unexpected results.  This code now deals with
combined short options like most UNIX utilities do, that is, “-ab” is
equivalent to “-a b” if “-a” takes an argument and “-a -b” if it
doesn’t.  Furthermore, “-ab 2” also works as expected (“-a -b 2”).