Remove method invocation count check
Methods should be callable any number of times.  If we want such methods
for a build tool, then that build tool should subclass Ame::Method for
that case and add a new statement, like single_invocation to a subclass
of Ame::Class, or something like along those lines.  That’s for later.